feat: Refactor profile management with modular classes
