+++ frontend/src/ts/profile/index.ts
@@ -1,0 +1,7 @@
+// Export all profile-related modules
+export { Profile } from "./Profile.js";
+export { UserCache, type CachedUser } from "./UserCache.js";
+export { ProfileStats } from "./ProfileStats.js";
+export { GameHistory } from "./GameHistory.js";
+export { FriendsList } from "./FriendsList.js";
+export { ProfileData } from "./ProfileData.js";
